requirements.txt
scikit_image==0.17.2
scandir==1.10.0
opencv_python==4.2.0.34
matplotlib==3.2.1
tqdm==4.46.1
tensorflow_gpu==2.2.0
Keras==2.3.1
face_recognition==1.3.0
pandas==1.0.4
dlib==19.20.0
numpy==1.18.5
Pillow==7.1.2
pygame==1.9.6
scikit_learn==0.23.1
skimage==0.0
tensorflow==2.2.0
torch==1.5.1
torchvision==0.6.1
